requirements.txt
python-telegram-bot>=20.0,<21.0
httpx>=0.23.0
aiosqlite>=0.19.0
ldap3>=2.9
imapclient>=2.3.1
poplib3>=1.0.3
email-validator>=2.0.0
